web/[Medium] OmniWatch/challenge/controller/requirements.txt
Flask==3.0.2
bcrypt==4.1.2
mysql-connector-python==8.3.0
python-dotenv==1.0.0
PyJWT==2.7.0
selenium==4.17.2 
schedule==1.2.1 